% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolationForest.R
\name{iForest}
\alias{iForest}
\title{iForest}
\usage{
<<<<<<< HEAD
iForest(X, nt = 100, phi = 256, seed = 1234, multicore = FALSE)
=======
iForest(X, nt = 100, phi = 256, seed = 1234, multicore = FALSE,
  replace_missing = TRUE, sentinel = -9999999999)
>>>>>>> 05649892b210a31e5578dba0d9d3352d82066858
}
\arguments{
\item{X}{a matrix or data.frame of numeric or factors values}

\item{nt}{the number of trees in the ensemble}

\item{phi}{the number of samples to draw without replacement to construct each tree}

\item{seed}{random seed to ensure creation of reproducible foresets}

\item{replace_missing}{if TRUE, replaces missing factor levels with "." and missing
numeric values with the \code{sentinel} argument}

\item{sentinel}{value to use as stand-in for missing numeric values}

\item{mulicore}{fit trees using the parallel package}
}
\value{
an \code{iForest} object
}
\description{
Build an Isolation Forest of completely random trees
}
\details{
An Isolation Forest is an unsupervised anomaly detection algorithm. The requested
number of trees, \code{nt}, are built completely at random on a subsample of size \code{phi}.
At each node a random variable is selected. A random split is chosen from the range of that
variable. A random sample of factor levels are chosen in the case the variable is a factor.

Records from \code{X} are then filtered based on the split criterion and the tree building
begins again on the left and right subsets of the data. Tree building terminates when the
maximum depth of the tree is reached or there are 1 or fewer observations in the filtered
subset.
}
\references{
F. T. Liu, K. M. Ting, Z.-H. Zhou, "Isolation-based anomaly detection",
\emph{ACM Trans. Knowl. Discov. Data}, vol. 6, no. 1, pp. 3:1-3:39, Mar. 2012.
}
