% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isofor_demo}
\alias{isofor_demo}
\title{Isolation Forest Demo
Shiny app demonstrating isolation forest parameters and
their affect on the anomaly score.}
\usage{
isofor_demo()
}
\description{
Isolation Forest Demo
Shiny app demonstrating isolation forest parameters and
their affect on the anomaly score.
}
