% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{isofor_to_sas}
\alias{isofor_to_sas}
\title{Export isolation forest to SAS code}
\usage{
isofor_to_sas(mod, pfx = "iso", nt = NULL)
}
\arguments{
\item{mod}{an isolation forest model}

\item{pfx}{the prefix for each isolation forest tree}

\item{nt}{optional number of trees to output. Defaults to all trees.}
}
\description{
Export isolation forest to SAS code
}
