% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolationForest.R
\name{print.iForest}
\alias{print.iForest}
\title{Printing iForest Objects}
\usage{
print.iForest(x, ...)
}
\arguments{
\item{x}{object of class iForest.}

\item{...}{ignored}
}
\description{
Printing iForest Objects
}
