% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolationForest.R
\name{iForest}
\alias{iForest}
\title{iForest}
\usage{
iForest(X, nt = 100, phi = 256)
}
\arguments{
\item{X}{a matrix or data.frame of numeric or factors values}

\item{nt}{the number of trees in the ensemble}

\item{phi}{the number of samples to draw without replacement to construct each tree}
}
\value{
an \code{iForest} object
}
\description{
Build an Isolation Forest of completly random trees
}
\details{
An Isolation Forest is an unsupervised anomaly detection algorithm. The requested
number of trees, \code{nt}, are built completely at random on a subsample of size \code{phi}.
At each node a random variable is selected. A random split is chosen from the range of that
variable. A random sample of factor levels are chosen in the case the variable is a factor.

Records from \code{X} are then filtered based on the split criterion and the tree building
begins again on the left and right subsets of the data. Tree building terminates when the
maximum depth of the tree is reached or there are 1 or fewer observations in the filtered
subset.
}

