% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.iForest}
\alias{predict.iForest}
\title{predcit.iForest}
\usage{
\method{predict}{iForest}(object, newdata, ..., type = c("score", "nodes",
  "terminal"))
}
\arguments{
\item{object}{an \code{iForest} object}

\item{newdata}{a dataset to predict}
}
\description{
predict.iForest is a method of the predict generic function.
}

