% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.iForest}
\alias{predict.iForest}
\title{predcit.iForest}
\usage{
\method{predict}{iForest}(object, newdata, ..., n.cores = 1, nodes = FALSE,
  sparse = FALSE)
}
\arguments{
\item{object}{an \code{iForest} object}

\item{newdata}{a dataset to predict}

\item{n.cores}{number of cores to use for prediction of anomaly score. Must
be compiled with openmp. Defaults to 1.}

\item{nodes}{if true return nobs x ntrees dim matrix with terminal node ids}

\item{sparse}{if true return sparse Matrix of dimension nobs x nTerminalNodes.
Each column represents a terminal node. There are as many ones in each row
as there are trees in the forest. Each observation can only belong to one
terminal node per tree. Useful for further modeling or to identify predictive
interactions.}
}
\description{
return predictions of various types for the isolation forest
object
}
\details{
By default the predict function returns an anomaly score. The
anomaly score is a [0,1] scaled measure of isolation. Higher scores
correspond to more isolated observations. If sparse or nodes are set to TRUE,
a matrix of the requested type is returned.
}
